more UI tweaks, wind direction!

+++ NiceWeather/Model/CurrentWeather.swift
@@ -48,7 +48,7 @@ struct CurrentWeather: Codable {
     
     struct Wind: Codable {
         let speed: Double
-        let deg: Int?
+        let deg: Double?
     }
     
     struct Sys: Codable {}
